Replace the tabbed resolve dialog with two single-purpose popups: Merge into Existing opens a merge-only dialog prefilled and searched with the unresolved name, and Create New Target opens a create-only dialog, in both the Target Management and Merges views
